Inline ProjectButton types
