+++ apps/frontend/app/components/ProjectButton/index.tsx
@@ -1,10 +1,18 @@
-import React from 'react';
-import { Appearance, StyleSheet, Text, TouchableOpacity } from 'react-native';
+import React, { ReactNode } from 'react';
+import { Appearance, StyleProp, StyleSheet, Text, TouchableOpacity, ViewStyle } from 'react-native';
 import { useSelector } from 'react-redux';
 import { RootState } from '@/redux/reducer';
 import { myContrastColor } from '@/helper/ColorHelper';
 import { darkTheme, lightTheme } from '@/styles/themes';
-import { ProjectButtonProps } from './types';
+
+export interface ProjectButtonProps {
+	text: string;
+	onPress?: () => void;
+	iconLeft?: ReactNode;
+	iconRight?: ReactNode;
+	backgroundColor?: string;
+	style?: StyleProp<ViewStyle>;
+}
 
 const ProjectButton: React.FC<ProjectButtonProps> = ({ text, onPress, iconLeft, iconRight, backgroundColor, style }) => {
 	const { primaryColor, selectedTheme } = useSelector((state: RootState) => state.settings);
